server/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymongo
google-generativeai